move GetDataChannel to private function in root webrtc pkg
